requirements.txt
biopython == 1.79
cobra == 0.29.0
joblib == 1.4.2
matplotlib == 3.5.1
numpy == 1.23.5
pandas == 2.2.2
pickleshare ~= 0.7
scikit-learn == 1.2.2
scipy == 1.9.3
jinja2 == 3.0.3
openpyxl == 3.1.0
    
# numpy#~=1.23
# pandas~=2.2.2 # 1.2.4
# openpyxl~=3.0.9
# cobra>=0.22
# jinja2<3.1.0
# #escher
# biopython==1.79
# matplotlib
# pip#==19.2.3
# bump2version==0.5.11
# wheel==0.33.6
# watchdog==0.9.0
# flake8==3.7.8
# tox==3.14.0
# coverage==4.5.4
# Sphinx==1.8.5
# sphinx_rtd_theme==1.0.0
# twine==1.14.0
# pytest==6.2.4
# black==21.7b0
